refactor: modify class be base on tradeoff
