+++ src/main/java/com/book_review/chap_02_object_oriented_programming/DiscountPolicyImpl/AmountDefaultDiscountPolicy.java
@@ -1,20 +1,20 @@
 package com.book_review.chap_02_object_oriented_programming.DiscountPolicyImpl;
 
 import com.book_review.chap_02_object_oriented_programming.DiscountCondition;
-import com.book_review.chap_02_object_oriented_programming.DiscountPolicy;
+import com.book_review.chap_02_object_oriented_programming.DefaultDiscountPolicy;
 import com.book_review.chap_02_object_oriented_programming.Money;
 import com.book_review.chap_02_object_oriented_programming.Screening;
 
-public class PercentDiscountPolicy extends DiscountPolicy {
-    private double percent;
+public class AmountDefaultDiscountPolicy extends DefaultDiscountPolicy {
+    private Money discountAmount;
 
-    public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
+    public AmountDefaultDiscountPolicy(Money discountAmount, DiscountCondition ... conditions) {
         super(conditions);
-        this.percent = percent;
+        this.discountAmount = discountAmount;
     }
 
     @Override
     protected Money getDiscountAmount(Screening screening) {
-        return screening.getMovieFee().times(percent);
+        return discountAmount;
     }
 }
