+++ src/main/java/com/book_review/chap_02_object_oriented_programming/DiscountPolicyImpl/NonDefaultDiscountPolicy.java
@@ -1,12 +1,13 @@
 package com.book_review.chap_02_object_oriented_programming.DiscountPolicyImpl;
 
+import com.book_review.chap_02_object_oriented_programming.DefaultDiscountPolicy;
 import com.book_review.chap_02_object_oriented_programming.DiscountPolicy;
 import com.book_review.chap_02_object_oriented_programming.Money;
 import com.book_review.chap_02_object_oriented_programming.Screening;
 
-public class NonDiscountPolicy extends DiscountPolicy {
+public class NonDefaultDiscountPolicy implements DiscountPolicy {
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
