+++ src/main/java/com/book_review/chap_02_object_oriented_programming/DiscountPolicyImpl/PercentDefaultDiscountPolicy.java
@@ -1,20 +1,20 @@
 package com.book_review.chap_02_object_oriented_programming.DiscountPolicyImpl;
 
 import com.book_review.chap_02_object_oriented_programming.DiscountCondition;
-import com.book_review.chap_02_object_oriented_programming.DiscountPolicy;
+import com.book_review.chap_02_object_oriented_programming.DefaultDiscountPolicy;
 import com.book_review.chap_02_object_oriented_programming.Money;
 import com.book_review.chap_02_object_oriented_programming.Screening;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
-    private Money discountAmount;
+public class PercentDefaultDiscountPolicy extends DefaultDiscountPolicy {
+    private double percent;
 
-    public AmountDiscountPolicy(Money discountAmount, DiscountCondition ... conditions) {
+    public PercentDefaultDiscountPolicy(double percent, DiscountCondition... conditions) {
         super(conditions);
-        this.discountAmount = discountAmount;
+        this.percent = percent;
     }
 
     @Override
     protected Money getDiscountAmount(Screening screening) {
-        return discountAmount;
+        return screening.getMovieFee().times(percent);
     }
 }
