+++ src/main/java/com/book_review/chap_02_object_oriented_programming/Movie.java
@@ -22,4 +22,8 @@ public class Movie {
     public Money calculateMovieFee(Screening screening) {
         return fee.minus(discountPolicy.calculateDiscountAmount(screening));
     }
+
+    public void changeDiscountPolicy(DiscountPolicy discountPolicy) {
+        this.discountPolicy = discountPolicy;
+    }
 }
